requirements.txt
django 
djangorestframework 
djangorestframework-simplejwt 
twilio 
python-dotenv
django-cors-headers
Pillow
razorpay